app/requirements.txt
altgraph==0.17
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.2
decorator==4.4.2
fake-useragent==0.1.11
flickrapi==2.4.0
future==0.18.2
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
joblib==0.15.1
lxml==4.5.1
macholib==1.14
moviepy==1.0.3
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
pathlib==1.0.1
pefile==2019.4.18
Pillow==7.1.2
proglog==0.1.9
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
tqdm==4.46.1
urllib3==1.25.9
websocket-client==0.48.0
